Update plan.md
I added several passages to round out our coverage of the Gospel of John. Note that at two points I'm proposing covering two passages on the same day.

We obviously need to pare down this list. We will not be able to cover it all, but I think it gives a good overview of John's Gospel using highly engaging passages.